+++ packages/specs/src/types/token.types.ts
@@ -174,7 +174,7 @@ export interface TypographyTokens {
   "text-sm--line-height": number; // 20
   "text-base--line-height": number; // 24
   "text-lg--line-height": number; // 28
-  "text-xl--line-height": number; // 30
+  "text-xl--line-height": number; // 28
   "text-2xl--line-height": number; // 32
   "text-3xl--line-height": number; // 36
   "text-4xl--line-height": number; // 40